+++ CHANGELOG.md
@@ -22,6 +22,7 @@ IMPROVEMENTS
 * [stake] keeper always loads the store (instead passing around which doesn't really boost efficiency)
 * [stake] edit-validator changes now can use the keyword [do-not-modify] to not modify unspecified `--flag` (aka won't set them to `""` value)
 * [types] added common tag constants
+* [stake] offload more generic functionality from the handler into the keeper
 
 FIXES
 * [cli] fixed cli-bash tests